좌표 정보 전달 방식 변경
